requirements.txt
IMAPClient==3.0.1
python-dotenv==1.2.1
